socrates/requirements.txt
grpcio-tools==1.33.2
allennlp==1.2.1
allennlp-models==1.2.1